Fixed offset for timing for forecasting
In the original code it used the pd.timedelta function in order to offset the time by one unit.  However, when created aggregations by week utilizing the Pandas Grouper function it creates a data type called 'W-SUN'.  This throws an error and won't allow the forecast to work.  This new code ought to read in the data type and offset it by one unit of the data type and then continue the forecast.  Here is a link to stackoverflow detailing it.

https://stackoverflow.com/questions/27368265/convert-freq-string-to-dateoffset-in-pandas